tools/top100-scraper/requirements.txt
anthropic>=0.40.0
pyautogui>=0.9.54
pygetwindow>=0.0.9
Pillow>=10.0.0
